// types/playerType.ts
export type PlayerType = {
  id: number;
  player_id: string;
  name: string;
  role: string;
  country: string;
  short_name: string;
  team_name: string;
  team_logo: string;
  team_short_name: string;
  event_total_points: number;
  event_player_credit: number;
  team_id: number;
  is_playing: boolean;
  player_stats_available: boolean;
};